mistral/pyproject.toml
[project]
name = "mistral"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mistralai>=1.5.1",
    "python-dotenv>=1.0.1",
    "tqdm>=4.67.1",
]
